// src/api/model/investmentProductGet.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InvestmentProductGet { 
    /**
     * Id of the created resource
     */
    readonly id: string;
    name: string;
    rate: number;
    /**
     * Type of the investment
     */
    type: InvestmentProductGet.TypeEnum;
}
export namespace InvestmentProductGet {
    export type TypeEnum = 'Action' | 'ETF' | 'Fund' | 'Bond' | 'Crypto' | 'RawMaterials';
    export const TypeEnum = {
        Action: 'Action' as TypeEnum,
        Etf: 'ETF' as TypeEnum,
        Fund: 'Fund' as TypeEnum,
        Bond: 'Bond' as TypeEnum,
        Crypto: 'Crypto' as TypeEnum,
        RawMaterials: 'RawMaterials' as TypeEnum
    };
}